ar_get_product_id: support older SST devices
Some AR carts and the like are manufactured with these older devices -
or sometimes one old and one new. Trying to detect using the JEDEC
command just ends up returning the Flash content for that address.

Now we try the proprietary command and see if we did get an SST device.
(This won't work if someone writes the first word of Flash to 0xbfbf,
so, like, don't do that?)